requirements.txt
sounddevice
python-vlc